Create new framework for recorders which are only included in iOS
