+++ Research/Research/RSDFileResultUtility.swift
@@ -1,8 +1,8 @@
 //
-//  RSDDataLogger.swift
+//  RSDFileResultUtility.swift
 //  Research
 //
-//  Copyright © 2017 Sage Bionetworks. All rights reserved.
+//  Copyright © 2017-2019 Sage Bionetworks. All rights reserved.
 //
 // Redistribution and use in source and binary forms, with or without modification,
 // are permitted provided that the following conditions are met:
@@ -33,69 +33,9 @@
 
 import Foundation
 
-/// `RSDDataLogger` is used to write data samples using a custom encoding to a logging file.
-/// - note: This class does **not** use a serial queue to process the samples. It is assumed that the
-/// recorder that is using this file will handle that implementation.
-open class RSDDataLogger {
-    
-    /// A unique identifier for the logger.
-    public let identifier: String
-    
-    /// The url to the file.
-    public let url: URL
-    
-    /// Open file handle for writing to the logger.
-    private let fileHandle: FileHandle
-    
-    /// Number of samples written to the file.
-    public private(set) var sampleCount: Int = 0
-    
-    /// The content type of the data file (if known).
-    open var contentType: String? {
-        return nil
-    }
-    
-    /// Default initializer. The initializer will automatically open the file and write the
-    /// initial data (if any).
-    ///
-    /// - parameters:
-    ///     - identifier: A unique identifier for the logger.
-    ///     - url: The url to the file.
-    ///     - initialData: The initial data to write to the file on opening.
-    public init(identifier: String, url: URL, initialData: Data?) throws {
-        self.identifier = identifier
-        self.url = url
-        
-        let data = initialData ?? Data()
-        try data.write(to: url)
-        
-        self.fileHandle = try FileHandle(forWritingTo: url)
-    }
-    
-    /// Write data to the logger.
-    /// - parameter data: The data to add to the logging file.
-    /// - throws: Error if writing the data fails because the wasn't enough memory on the device.
-    open func write(_ data: Data) throws {
-        try RSDExceptionHandler.try {
-            self.fileHandle.seekToEndOfFile()
-            self.fileHandle.write(data)
-        }
-        sampleCount += 1
-    }
-    
-    /// Close the file. This will write the end tag for the root element and then close the file handle.
-    /// If there is an error thrown by writing the closing tag, then the file handle will be closed and
-    /// the error will be rethrown.
-    ///
-    /// - throws: Error thrown when attempting to write the closing tag.
-    open func close() throws {
-        self.fileHandle.closeFile()
-    }
-}
-
 /// `RSDFileResultUtility` is a utility for naming temporary files used to save task results.
 public class RSDFileResultUtility {
-
+    
     /// This utility will create a filename scrubbing the identifier string of any characters that
     /// are not alpha-numeric, dash, or underscore characters. Additionally, this utility will
     /// replace '.' and whitespace characters with an underscore. Finally, the string will be
@@ -136,7 +76,7 @@ public class RSDFileResultUtility {
     /// - returns: Scrubbed URL for the given identifier.
     /// - throws: An exception if the file directory cannot be created.
     public static func createFileURL(identifier: String, ext: String, outputDirectory: URL, shouldDeletePrevious: Bool = false) throws -> URL {
-
+        
         // create the directory if needed
         try FileManager.default.createDirectory(at: outputDirectory, withIntermediateDirectories: true, attributes: nil)
         
@@ -179,3 +119,4 @@ public class RSDFileResultUtility {
         return RSDFileManifest(filename: filename, timestamp: Date(), contentType: "application/json")
     }
 }
+
