+++ Research/ResearchRecorders/Permissions/RSDMotionAuthorization.swift
