+++ Research/ResearchRecorders/iOS/RSDDistanceRecorder.swift
@@ -618,7 +618,8 @@ extension CLLocation {
 
 // Documentation and Tests
 
-extension RSDDistanceRecord : RSDDocumentableCodableObject {
+// TODO: syoung 09/27/2019 Look into whether or not there is a simple way to use the Documentable protocols in other frameworks.
+extension RSDDistanceRecord { //: RSDDocumentableCodableObject {
     
     public static func codingKeys() -> [CodingKey] {
         return CodingKeys.allCases
