+++ Research/ResearchRecorders/iOS/RSDMotionRecorder.swift
@@ -568,7 +568,8 @@ extension CMMagnetometerData : RSDVectorData {
 
 // Documentation and Tests
 
-extension RSDMotionRecord : RSDDocumentableCodableObject {
+// TODO: syoung 09/27/2019 Look into whether or not there is a simple way to use the Documentable protocols in other frameworks.
+extension RSDMotionRecord { //}: RSDDocumentableCodableObject {
     
     public static func codingKeys() -> [CodingKey] {
         return CodingKeys.allCases
