+++ Research/ResearchRecordersTests/CodableMotionRecorderTests.swift
@@ -33,6 +33,7 @@
 
 import XCTest
 @testable import Research
+@testable import ResearchRecorders
 
 class CodableMotionRecorderTests: XCTestCase {
